Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    int num1,num2,num3;
+    float num1,num2,num3;
     char oprtr;
-    scanf("%d %d %c",&num1,&num2,&oprtr);
+    scanf("%f %f %c",&num1,&num2,&oprtr);
     if (oprtr=='+') {
         num3=num1+num2;
 
@@ -14,6 +14,6 @@ int main() {
     } else if (oprtr=='/') {
         num3=num1/num2;
     } 
-    printf("%d",num3);
+    printf("%.0f",num3);
     return 0;
 }
